Added string values.  They can be concatenated.
Quote quotes by doubling them up.

TIME$ is no longer a special case for EQUS.

Introduces a Value type that can be a string or a double.  This uses its
own string type.  Alternatives were to not use a union, or to bring in a
big dependency like boost to handle complex unions.