Failed recovery from undefined symbol in function in list
When an undefined symbol is evaluated beebasm skips the rest of the
expression and throws an exception.  Previously, a comma was
unequivocally the end of an expression but this is not true with
multi-parameter functions.  LineParser::SkipExpression needs to keep
track of the bracket count to handle commas correctly.  This only caused
problems when an undefined symbol was surrounded with brackets (like in
a function call), and there were multiple expressions in a comma-separated
list, like in an EQUB command.