Fix ASSERT directive to work correctly
Casting the internal double result to an unsigned int is undefined
behavior. This becomes an issue with the use of -1 to represent a True
result. The cast to unsigned int clamps -1 to 0 which is interpreted as
the assert expression being False. The fix is to switch to signed int.

Cleaned up the assertdemo.6502 to show obviously passing and failing
expressions with the ASSERT directive.